pyproject.toml
[project]
name = "lakehouse-elt"
version = "0.1.0"
description = ""
authors = [
    {name = "Jeova BL",email = "jeovab115@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "ruff (>=0.9.7,<0.10.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "dbt-postgres (>=1.9.0,<2.0.0)",
    "polars (>=1.23.0,<2.0.0)",
    "scrapy (>=2.12.0,<3.0.0)",
    "scrapy-proxy-pool (>=0.1.9,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'